requirements.txt
pandas == 2.1.4
numpy == 1.26.2
matplotlib == 3.8.2
seaborn == 0.13.0
streamlit == 1.29.0
Pillow == 10.1.0